Refactor: Step별 css파일 (step3 화면 크기 오류)
