+++ src/App.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Header from "./components/Header/Header";
-import TimetableResult from "./components/TimetableResult/TimetableResult";
-import TimetableForm from "./components/TimetableFrom/TimetableForm";
+import Header from "./Header/Header";
+import TimetableResult from "./TimetableResult/TimetableResult";
+import TimetableForm from "./TimetableFrom/TimetableForm";
 
 function App() {
   return (
