+++ src/TimetableFrom/ProgressBar/ProgressBar.css
