+++ src/TimetableFrom/Step1YearSemester/Step1.css
@@ -1,0 +1,11 @@
+.year-buttons {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 0.75rem;
+}
+
+.semester-buttons {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 0.75rem;
+}
