+++ src/TimetableFrom/Step2Department/Step2.css
@@ -1,0 +1,101 @@
+.department-container {
+  width: 100%;
+  background-color: white;
+  border: 1px solid #d1d5db;
+  border-radius: 0.375rem;
+  overflow: hidden;
+  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
+}
+
+.department-section {
+  padding: 1rem;
+}
+
+.department-section-title {
+  font-size: 1.125rem;
+  font-weight: 500;
+  color: #1f2937;
+  margin-bottom: 1rem;
+}
+
+.college-list {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.college-button {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.75rem;
+  border-radius: 0.25rem;
+  border: 1px solid #e5e7eb;
+  transition: background-color 0.2s, border-color 0.2s;
+}
+
+.college-button:hover {
+  background-color: #f9fafb;
+}
+
+.college-icon {
+  height: 1.25rem;
+  width: 1.25rem;
+  color: #9ca3af;
+}
+
+.back-button {
+  display: flex;
+  align-items: center;
+  font-size: 0.875rem;
+  color: #2563eb;
+  margin-bottom: 1rem;
+}
+
+.back-button:hover {
+  color: #1d4ed8;
+}
+
+.back-icon {
+  height: 1rem;
+  width: 1rem;
+  margin-right: 0.25rem;
+}
+
+.department-list {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  max-height: 15rem;
+  overflow-y: auto;
+}
+
+.department-button {
+  width: 100%;
+  text-align: left;
+  padding: 0.75rem;
+  border-radius: 0.25rem;
+  transition: all 0.2s;
+}
+
+.department-button-selected {
+  background-color: #2563eb;
+  color: white;
+}
+
+.department-button-unselected {
+  background-color: white;
+  color: #1f2937;
+  border: 1px solid #e5e7eb;
+}
+
+.department-button-unselected:hover {
+  background-color: #f9fafb;
+}
+
+.navigation-buttons {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 2rem;
+}
