+++ src/TimetableFrom/Step3Credits/Step3.css
@@ -1,0 +1,33 @@
+.credits-grid {
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  gap: 0.2rem;
+}
+
+.credit-button {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0.5rem 0;
+  border-radius: 0.375rem;
+  font-weight: 500;
+  transition: all 0.2s;
+}
+
+.credits-summary {
+  margin-bottom: 1.5rem;
+  padding: 1rem;
+  background-color: #ebf5ff;
+  border: 1px solid #bfdbfe;
+  border-radius: 0.375rem;
+}
+
+.summary-text {
+  color: #1e40af;
+  font-weight: 500;
+}
+
+.warning-text {
+  color: #f97316;
+  margin-left: 0.5rem;
+}
