+++ src/TimetableFrom/TimetableFormStyles.css
@@ -1,0 +1,189 @@
+/*TimetableForm 전체 스타일*/
+
+/* 슬라이드 애니메이션 - 왼쪽에서 오른쪽으로 */
+.slide-left {
+  animation: slideLeft 0.5s ease-out forwards;
+}
+
+@keyframes slideLeft {
+  0% {
+    transform: translateX(100%);
+    opacity: 0;
+  }
+  100% {
+    transform: translateX(0);
+    opacity: 1;
+  }
+}
+
+/* 슬라이드 애니메이션 - 오른쪽에서 왼쪽으로 */
+.slide-right {
+  animation: slideRight 0.5s ease-out forwards;
+}
+
+@keyframes slideRight {
+  0% {
+    transform: translateX(-100%);
+    opacity: 0;
+  }
+  100% {
+    transform: translateX(0);
+    opacity: 1;
+  }
+}
+
+/* 단계 전환 시 페이드 효과 */
+.wizard-container {
+  position: relative;
+  min-height: 400px;
+}
+
+/* 버튼 호버 효과 */
+.btn-hover-effect {
+  transition: all 0.2s ease;
+}
+
+.btn-hover-effect:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+/* 체크박스 선택 효과 */
+.checkbox-animation {
+  transition: all 0.2s ease;
+}
+
+.checkbox-animation:checked {
+  transform: scale(1.1);
+}
+
+/* 진행 바 애니메이션 */
+.progress-bar-fill {
+  transition: width 0.5s ease-out;
+}
+
+/* 반응형 디자인을 위한 미디어 쿼리 */
+@media (max-width: 768px) {
+  .wizard-container {
+    min-height: 500px;
+  }
+}
+
+.timetable-container {
+  background-color: #f9fafb;
+  min-height: 100vh;
+  padding: 1.5rem;
+}
+
+.timetable-inner {
+  max-width: 48rem;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+/* 헤더 스타일 */
+.timetable-header {
+  text-align: center;
+  margin-bottom: 2rem;
+}
+
+.timetable-title {
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: #1f2937;
+}
+
+.timetable-subtitle {
+  color: #4b5563;
+  margin-top: 0.5rem;
+}
+
+/* 컨텐츠 카드 스타일 */
+.timetable-card {
+  margin-top: 2rem;
+  background-color: white;
+  padding: 2rem;
+  border-radius: 0.5rem;
+  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1),
+    0 2px 4px -1px rgba(0, 0, 0, 0.06);
+}
+
+.timetable-content {
+  transition-property: all;
+  transition-duration: 500ms;
+}
+
+.step-container {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+.step-title {
+  font-size: 1.25rem;
+  font-weight: 600;
+  text-align: center;
+  margin-bottom: 1.5rem;
+}
+
+.step-section {
+  margin-bottom: 1.5rem;
+}
+
+.step-label {
+  display: block;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: #374151;
+  margin-bottom: 0.75rem;
+}
+
+.step-footer {
+  display: flex;
+  justify-content: flex-end;
+  margin-top: 2rem;
+}
+
+/* 버튼 스타일 */
+.button-active {
+  padding: 0.5rem 1rem;
+  background-color: #2563eb;
+  color: white;
+  border-radius: 0.375rem;
+  font-weight: 500;
+}
+
+.button-inactive {
+  padding: 0.5rem 1rem;
+  background-color: #f3f4f6;
+  color: #374151;
+  border-radius: 0.375rem;
+  font-weight: 500;
+}
+
+.button-inactive:hover {
+  background-color: #e5e7eb;
+}
+
+.next-button {
+  padding: 0.75rem 1.5rem;
+  background-color: #2563eb;
+  color: white;
+  border-radius: 0.375rem;
+  font-weight: 500;
+}
+
+.next-button:hover {
+  background-color: #1d4ed8;
+}
+
+.prev-button {
+  padding: 0.75rem 1.5rem;
+  background-color: #d1d5db;
+  color: #374151;
+  border-radius: 0.375rem;
+}
+
+.prev-button:hover {
+  background-color: #9ca3af;
+}
