--- src/components/TimetableFrom/TimetableFormStyles.css
@@ -1,439 +1,0 @@
-/*TimetableForm 전체 스타일*/
-
-/* 슬라이드 애니메이션 - 왼쪽에서 오른쪽으로 */
-.slide-left {
-  animation: slideLeft 0.5s ease-out forwards;
-}
-
-@keyframes slideLeft {
-  0% {
-    transform: translateX(100%);
-    opacity: 0;
-  }
-  100% {
-    transform: translateX(0);
-    opacity: 1;
-  }
-}
-
-/* 슬라이드 애니메이션 - 오른쪽에서 왼쪽으로 */
-.slide-right {
-  animation: slideRight 0.5s ease-out forwards;
-}
-
-@keyframes slideRight {
-  0% {
-    transform: translateX(-100%);
-    opacity: 0;
-  }
-  100% {
-    transform: translateX(0);
-    opacity: 1;
-  }
-}
-
-/* 단계 전환 시 페이드 효과 */
-.wizard-container {
-  position: relative;
-  min-height: 400px;
-}
-
-/* 버튼 호버 효과 */
-.btn-hover-effect {
-  transition: all 0.2s ease;
-}
-
-.btn-hover-effect:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-/* 체크박스 선택 효과 */
-.checkbox-animation {
-  transition: all 0.2s ease;
-}
-
-.checkbox-animation:checked {
-  transform: scale(1.1);
-}
-
-/* 진행 바 애니메이션 */
-.progress-bar-fill {
-  transition: width 0.5s ease-out;
-}
-
-/* 반응형 디자인을 위한 미디어 쿼리 */
-@media (max-width: 768px) {
-  .wizard-container {
-    min-height: 500px;
-  }
-}
-
-.timetable-container {
-  background-color: #f9fafb;
-  min-height: 100vh;
-  padding: 1.5rem;
-}
-
-.timetable-inner {
-  max-width: 48rem;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-/* 헤더 스타일 */
-.timetable-header {
-  text-align: center;
-  margin-bottom: 2rem;
-}
-
-.timetable-title {
-  font-size: 1.5rem;
-  font-weight: 700;
-  color: #1f2937;
-}
-
-.timetable-subtitle {
-  color: #4b5563;
-  margin-top: 0.5rem;
-}
-
-/* 컨텐츠 카드 스타일 */
-.timetable-card {
-  margin-top: 2rem;
-  background-color: white;
-  padding: 2rem;
-  border-radius: 0.5rem;
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1),
-    0 2px 4px -1px rgba(0, 0, 0, 0.06);
-}
-
-.timetable-content {
-  transition-property: all;
-  transition-duration: 500ms;
-}
-
-.step-container {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.step-title {
-  font-size: 1.25rem;
-  font-weight: 600;
-  text-align: center;
-  margin-bottom: 1.5rem;
-}
-
-.step-section {
-  margin-bottom: 1.5rem;
-}
-
-.step-label {
-  display: block;
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: #374151;
-  margin-bottom: 0.75rem;
-}
-
-.step-footer {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 2rem;
-}
-
-/* 버튼 스타일 */
-.button-active {
-  padding: 0.5rem 1rem;
-  background-color: #2563eb;
-  color: white;
-  border-radius: 0.375rem;
-  font-weight: 500;
-}
-
-.button-inactive {
-  padding: 0.5rem 1rem;
-  background-color: #f3f4f6;
-  color: #374151;
-  border-radius: 0.375rem;
-  font-weight: 500;
-}
-
-.button-inactive:hover {
-  background-color: #e5e7eb;
-}
-
-.next-button {
-  padding: 0.75rem 1.5rem;
-  background-color: #2563eb;
-  color: white;
-  border-radius: 0.375rem;
-  font-weight: 500;
-}
-
-.next-button:hover {
-  background-color: #1d4ed8;
-}
-
-.prev-button {
-  padding: 0.75rem 1.5rem;
-  background-color: #d1d5db;
-  color: #374151;
-  border-radius: 0.375rem;
-}
-
-.prev-button:hover {
-  background-color: #9ca3af;
-}
-
-/*Step1 버튼 스타일*/
-.year-buttons {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 0.75rem;
-}
-
-.semester-buttons {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 0.75rem;
-}
-
-/*Step2 */
-.department-container {
-  width: 100%;
-  background-color: white;
-  border: 1px solid #d1d5db;
-  border-radius: 0.375rem;
-  overflow: hidden;
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
-}
-
-.department-section {
-  padding: 1rem;
-}
-
-.department-section-title {
-  font-size: 1.125rem;
-  font-weight: 500;
-  color: #1f2937;
-  margin-bottom: 1rem;
-}
-
-.college-list {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.college-button {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.75rem;
-  border-radius: 0.25rem;
-  border: 1px solid #e5e7eb;
-  transition: background-color 0.2s, border-color 0.2s;
-}
-
-.college-button:hover {
-  background-color: #f9fafb;
-}
-
-.college-icon {
-  height: 1.25rem;
-  width: 1.25rem;
-  color: #9ca3af;
-}
-
-.back-button {
-  display: flex;
-  align-items: center;
-  font-size: 0.875rem;
-  color: #2563eb;
-  margin-bottom: 1rem;
-}
-
-.back-button:hover {
-  color: #1d4ed8;
-}
-
-.back-icon {
-  height: 1rem;
-  width: 1rem;
-  margin-right: 0.25rem;
-}
-
-.department-list {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  max-height: 15rem;
-  overflow-y: auto;
-}
-
-.department-button {
-  width: 100%;
-  text-align: left;
-  padding: 0.75rem;
-  border-radius: 0.25rem;
-  transition: all 0.2s;
-}
-
-.department-button-selected {
-  background-color: #2563eb;
-  color: white;
-}
-
-.department-button-unselected {
-  background-color: white;
-  color: #1f2937;
-  border: 1px solid #e5e7eb;
-}
-
-.department-button-unselected:hover {
-  background-color: #f9fafb;
-}
-
-.navigation-buttons {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 2rem;
-}
-
-/* Step3 - 학점 선택 스타일 */
-.credits-grid {
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  gap: 0.2rem;
-}
-
-.credit-button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 0.5rem 0;
-  border-radius: 0.375rem;
-  font-weight: 500;
-  transition: all 0.2s;
-}
-
-.credits-summary {
-  margin-bottom: 1.5rem;
-  padding: 1rem;
-  background-color: #ebf5ff;
-  border: 1px solid #bfdbfe;
-  border-radius: 0.375rem;
-}
-
-.summary-text {
-  color: #1e40af;
-  font-weight: 500;
-}
-
-.warning-text {
-  color: #f97316;
-  margin-left: 0.5rem;
-}
-
-/* 동적 버튼 텍스트 스타일 */
-.submit-button-text {
-  display: inline-block;
-}
-
-/* Step4 - 교양 영역 선택 스타일 */
-.areas-info {
-  font-size: 0.875rem;
-  color: #6b7280;
-  margin-bottom: 1rem;
-}
-
-.areas-grid {
-  display: grid;
-  grid-template-columns: repeat(1, 1fr);
-  gap: 0.75rem;
-}
-
-@media (min-width: 768px) {
-  .areas-grid {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-.area-button {
-  display: flex;
-  align-items: center;
-  padding: 0.75rem 1rem;
-  border-width: 1px;
-  border-radius: 0.375rem;
-  transition: all 0.2s;
-}
-
-.area-button-selected {
-  background-color: #ebf5ff;
-  border-color: #93c5fd;
-  color: #1d4ed8;
-}
-
-.area-button-unselected {
-  border-color: #d1d5db;
-  color: #374151;
-}
-
-.area-button-unselected:hover {
-  background-color: #f9fafb;
-}
-
-.checkbox {
-  width: 1.25rem;
-  height: 1.25rem;
-  border-width: 1px;
-  border-radius: 0.25rem;
-  margin-right: 0.75rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.checkbox-selected {
-  background-color: #2563eb;
-  border-color: #2563eb;
-}
-
-.checkbox-unselected {
-  border-color: #9ca3af;
-}
-
-.checkbox-icon {
-  width: 0.75rem;
-  height: 0.75rem;
-  color: white;
-}
-
-.area-name {
-  flex: 1;
-  text-align: left;
-}
-
-.tag {
-  margin-left: 0.5rem;
-  font-size: 0.75rem;
-  background-color: #f3f4f6;
-  padding: 0.25rem 0.5rem;
-  border-radius: 9999px;
-}
-
-.submit-button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-width: 8rem;
-}
-
-.loading-indicator {
-  opacity: 0.7;
-  cursor: not-allowed;
-}
